Alfred 49

+++ app/controllers/categories_controller.rb
@@ -18,6 +18,7 @@ class CategoriesController < ApplicationController
     @category=Category.new(category_params)
     cat_name=Category.find_by_name @category.name
     if cat_name.nil?
+      @category.status = 0
       @category.save
       redirect_to @category
     else
@@ -49,15 +50,29 @@ class CategoriesController < ApplicationController
    cat = Category.find_by(id: params[:category_id])
    unless cat.nil?
       name = cat.name
-      cat.destroy
+      if (cat.has_milestones?)
+        cat.status = 1
+        cat.save
+      else
+        cat.destroy
+      end
+
       flash.notice = "#{name} " + t('messages.delete.success')
     end
     redirect_to categories_path
   end
 
+  def activate
+
+    @cate = Category.find_by(id: params[:id] || params[:category_id])
+    @cate.status = 0
+    @cate.save!
+    redirect_to :back
+  end
+
   private
   def category_params
-    params.require(:category).permit(:name, :icon, :created_at, :updated_at,:doc_url,:is_feedback)
+    params.require(:category).permit(:name, :icon, :created_at, :updated_at,:is_feedback)
   end
